third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-history-back-after-pushState.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  window.onload = t.step_timeout(() => {
    let start_length = history.length;
    history.pushState(1, "", "pushState.html");
    assert_equals(history.length, start_length + 1);

    appHistory.onnavigate = t.step_func_done(e => {
      assert_equals(e.canRespond, true);
      assert_equals(e.hashChange, false);
      assert_equals(e.formData, null);
    });

    history.back();
  }, 0);
}, "history.back() fires the navigate event when reversing a pushState()");
</script>
